postgrado-ia-python/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and RAG
langchain==0.1.3
langchain-openai==0.0.5
openai==1.3.0
chromadb==0.4.18

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.1

# Database
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
requests==2.31.0
python-multipart==0.0.6

# Logging and Monitoring
python-json-logger==2.0.7

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
